Fix _atoi function to use double for result and include limits.h

+++ pointers_arrays_strings/100-atoi.c
@@ -1,5 +1,6 @@
 #include "main.h"
 #include <stdio.h>
+#include <limits.h>
 /**
  * _atoi - converts a string to an integer.
  * @s: the string to be converted
@@ -11,7 +12,7 @@ int _atoi(char *s)
 {
 	int i = 0;
 	int sign = 1;
-	int result = 0;
+	double result = 0;
 	int found_digit = 0;
 
 	while (s[i])
